WIP - collider

+++ Assets/Gridlike/Lib/Scripts/Models/Tile.cs
@@ -9,7 +9,7 @@ public class Tile {
 	public int id;
 	public int subId;
 
-	public int state1;
-	public int state2;
-	public int state3;
+	public float state1;
+	public float state2;
+	public float state3;
 }
